requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
pydantic
python-dateutil
beautifulsoup4
lxml
